Reduce fake server logging when running locally or in CI.
