+++ app/lib/fake/server/fake_server_entrypoint.dart
@@ -17,6 +17,8 @@ import 'package:pub_dev/fake/server/fake_storage_server.dart';
 import 'package:pub_dev/fake/server/local_server_state.dart';
 import 'package:pub_dev/frontend/static_files.dart';
 import 'package:pub_dev/shared/configuration.dart';
+import 'package:pub_dev/shared/env_config.dart';
+import 'package:pub_dev/shared/logging.dart';
 import 'package:pub_dev/task/cloudcompute/fakecloudcompute.dart';
 import 'package:pub_dev/tool/test_profile/import_source.dart';
 import 'package:pub_dev/tool/test_profile/importer.dart';
@@ -66,14 +68,18 @@ class FakeServerCommand extends Command {
     final dataFile = argResults!['data-file'] as String?;
     final watch = argResults!['watch'] == true;
 
-    Logger.root.onRecord.listen((r) {
-      print([
-        r.time.toIso8601String(),
-        r.toString(),
-        r.error,
-        r.stackTrace?.toString(),
-      ].where((e) => e != null).join(' '));
-    });
+    if (envConfig.isInsideCI) {
+      setupDebugEnvBasedLogging(extraDebug: 'fake_server pub.email');
+    } else {
+      Logger.root.onRecord.listen((r) {
+        print([
+          r.time.toIso8601String(),
+          r.toString(),
+          r.error,
+          r.stackTrace?.toString(),
+        ].where((e) => e != null).join(' '));
+      });
+    }
 
     final state = LocalServerState();
     if (dataFile != null) {
