+++ app/lib/shared/env_config.dart
@@ -45,6 +45,9 @@ class _EnvConfig {
   @visibleForTesting
   late final debug = Platform.environment['DEBUG'];
 
+  /// True, if the app (or test) is running inside CI.
+  bool get isInsideCI => Platform.environment.containsKey('CI');
+
   /// True, if running inside AppEngine.
   bool get isRunningInAppengine => _gaeService != null && _gaeVersion != null;
