+++ pkg/pub_integration/test/publisher_test.dart
@@ -31,7 +31,8 @@ void main() {
 
       Future<void> inviteCompleterFn() async {
         final pageUrl = await inviteUrlFuture.timeout(Duration(seconds: 30));
-        final pageUri = Uri.parse(pageUrl);
+        final pageUri = Uri.parse(
+            pageUrl.substring(pageUrl.indexOf('https://pub.dev/consent')));
         final consentId = pageUri.queryParameters['id'];
 
         // spoofed consent, trying to accept it with a different user
